backend/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
pymongo==4.5.0
pydantic==2.4.2
python-dotenv==1.0.0
openai>=1.0.0
pillow==10.0.1
numpy==1.26.0
torch==2.1.0
diffusers==0.23.1
transformers==4.34.1
celery==5.3.4
redis==5.0.1
requests==2.31.0
python-multipart==0.0.6
starlette==0.27.0
httpx==0.25.0
tenacity==8.2.3
pymilvus==2.3.1
faiss-cpu==1.7.4 